+++ __tests__/account.test.ts
@@ -3,13 +3,15 @@ import { isBN } from 'bn.js';
 import typedDataExample from '../__mocks__/typedDataExample.json';
 import { Account, Contract, DeployContractPayload, Provider, number, stark } from '../src';
 import { feeTransactionVersion } from '../src/utils/hash';
-import { toBN } from '../src/utils/number';
+import { hexToDecimalString, toBN } from '../src/utils/number';
+import { encodeShortString } from '../src/utils/shortString';
+import { randomAddress } from '../src/utils/stark';
 import {
   compiledErc20,
   compiledNamingContract,
   compiledStarknetId,
   compiledTestDapp,
-  getERC20DeployPayload,
+  erc20ClassHash,
   getTestAccount,
   getTestProvider,
 } from './fixtures';
@@ -24,25 +26,29 @@ describe('deploy and test Wallet', () => {
   beforeAll(async () => {
     expect(account).toBeInstanceOf(Account);
 
-    const erc20DeployPayload = getERC20DeployPayload(account.address);
+    const declareDeploy = await account.declareDeploy({
+      contract: compiledErc20,
+      classHash: '0x54328a1075b8820eb43caf0caa233923148c983742402dcfc38541dd843d01a',
+      constructorCalldata: [
+        encodeShortString('Token'),
+        encodeShortString('ERC20'),
+        account.address,
+      ],
+    });
 
-    const erc20Response = await provider.deployContract(erc20DeployPayload);
-
-    erc20Address = erc20Response.contract_address;
+    erc20Address = declareDeploy.deploy.contract_address;
     erc20 = new Contract(compiledErc20.abi, erc20Address, provider);
-
-    await provider.waitForTransaction(erc20Response.transaction_hash);
 
     const x = await erc20.balanceOf(account.address);
 
     expect(number.toBN(x[0].low).toString()).toStrictEqual(number.toBN(1000).toString());
 
-    const dappResponse = await provider.deployContract({
+    const dappResponse = await account.declareDeploy({
       contract: compiledTestDapp,
+      classHash: '0x04367b26fbb92235e8d1137d19c080e6e650a6889ded726d00658411cc1046f5',
     });
-    dapp = new Contract(compiledTestDapp.abi, dappResponse.contract_address!, provider);
 
-    await provider.waitForTransaction(dappResponse.transaction_hash);
+    dapp = new Contract(compiledTestDapp.abi, dappResponse.deploy.contract_address!, provider);
   });
 
   test('estimate fee', async () => {
@@ -70,7 +76,7 @@ describe('deploy and test Wallet', () => {
       calldata: [erc20.address, '10', '0'],
     });
 
-    await provider.waitForTransaction(transaction_hash);
+    await provider.waitForTransaction(transaction_hash, ['ACCEPTED_ON_L2']);
   });
 
   test('read balance of wallet after transfer', async () => {
@@ -109,7 +115,7 @@ describe('deploy and test Wallet', () => {
       },
     ]);
 
-    await provider.waitForTransaction(transaction_hash);
+    await provider.waitForTransaction(transaction_hash, ['ACCEPTED_ON_L2']);
 
     const response = await dapp.get_number(account.address);
     expect(toBN(response.number as string).toString()).toStrictEqual('57');
@@ -218,6 +224,60 @@ describe('deploy and test Wallet', () => {
 
       const name = await account.getStarkName(namingAddress);
       expect(name).toEqual('ben.stark');
+    });
+  });
+
+  describe('Declare and UDC Deploy Flow', () => {
+    test('ERC20 Declare', async () => {
+      const declareTx = await account.declare({
+        classHash: erc20ClassHash,
+        contract: compiledErc20,
+      });
+
+      await provider.waitForTransaction(declareTx.transaction_hash);
+
+      expect(declareTx).toHaveProperty('class_hash');
+      expect(hexToDecimalString(declareTx.class_hash)).toEqual(hexToDecimalString(erc20ClassHash));
+    });
+
+    test('UDC DeployContract', async () => {
+      const deployResponse = await account.deployContract({
+        classHash: erc20ClassHash,
+        constructorCalldata: [
+          encodeShortString('Token'),
+          encodeShortString('ERC20'),
+          account.address,
+        ],
+      });
+
+      expect(deployResponse.contract_address).toBeDefined();
+      expect(deployResponse.transaction_hash).toBeDefined();
+      expect(deployResponse.address).toBeDefined();
+      expect(deployResponse.deployer).toBeDefined();
+      expect(deployResponse.unique).toBeDefined();
+      expect(deployResponse.classHash).toBeDefined();
+      expect(deployResponse.calldata_len).toBeDefined();
+      expect(deployResponse.calldata).toBeDefined();
+      expect(deployResponse.salt).toBeDefined();
+    });
+
+    test('UDC Deploy', async () => {
+      const salt = randomAddress(); // use random salt
+
+      const deployment = await account.deploy({
+        classHash: erc20ClassHash,
+        constructorCalldata: [
+          encodeShortString('Token'),
+          encodeShortString('ERC20'),
+          account.address,
+        ],
+        salt,
+        unique: true, // Using true here so as not to clash with normal erc20 deploy in account and provider test
+      });
+
+      await provider.waitForTransaction(deployment.transaction_hash);
+
+      expect(deployment).toHaveProperty('transaction_hash');
     });
   });
 });