+++ __tests__/rpcProvider.test.ts
@@ -2,21 +2,19 @@ import { Account, GetBlockResponse, RpcProvider, ec } from '../src';
 import { StarknetChainId } from '../src/constants';
 import {
   compiledOpenZeppelinAccount,
+  describeIfNotDevnet,
   describeIfRpc,
   getTestAccount,
   getTestProvider,
 } from './fixtures';
 
 describeIfRpc('RPCProvider', () => {
-  let rpcProvider: RpcProvider;
+  const rpcProvider = getTestProvider() as RpcProvider;
+  const account = getTestAccount(rpcProvider);
   let accountPublicKey: string;
 
   beforeAll(async () => {
-    rpcProvider = getTestProvider() as RpcProvider;
-    const account = getTestAccount(rpcProvider);
-
     expect(account).toBeInstanceOf(Account);
-
     const accountKeyPair = ec.genKeyPair();
     accountPublicKey = ec.getStarkKey(accountKeyPair);
   });
@@ -26,11 +24,6 @@ describeIfRpc('RPCProvider', () => {
     expect([StarknetChainId.TESTNET2, StarknetChainId.MAINNET, StarknetChainId.TESTNET]).toContain(
       chainId
     );
-  });
-
-  test('getPendingTransactions', async () => {
-    const transactions = await rpcProvider.getPendingTransactions();
-    expect(Array.isArray(transactions)).toBe(true);
   });
 
   test('getTransactionCount', async () => {
@@ -52,8 +45,15 @@ describeIfRpc('RPCProvider', () => {
     expect(stateUpdate).toHaveProperty('state_diff');
   });
 
-  xtest('getProtocolVersion', async () => {
+  xtest('getProtocolVersion - pathfinder not implement', async () => {
     await rpcProvider.getProtocolVersion();
+  });
+
+  describeIfNotDevnet('devnet not implement', () => {
+    test('getPendingTransactions', async () => {
+      const transactions = await rpcProvider.getPendingTransactions();
+      expect(Array.isArray(transactions)).toBe(true);
+    });
   });
 
   describe('RPC methods', () => {
@@ -90,15 +90,18 @@ describeIfRpc('RPCProvider', () => {
       let transaction_hash;
 
       beforeAll(async () => {
-        ({ contract_address, transaction_hash } = await rpcProvider.deployContract({
+        const { deploy } = await account.declareDeploy({
           contract: compiledOpenZeppelinAccount,
+          classHash: '0x03fcbf77b28c96f4f2fb5bd2d176ab083a12a5e123adeb0de955d7ee228c9854',
           constructorCalldata: [accountPublicKey],
-          addressSalt: accountPublicKey,
-        }));
-        await rpcProvider.waitForTransaction(transaction_hash);
+          salt: accountPublicKey,
+        });
+
+        contract_address = deploy.contract_address;
+        transaction_hash = deploy.transaction_hash;
       });
 
-      test('deployContract result', () => {
+      test('declareDeploy()', () => {
         expect(contract_address).toBeTruthy();
         expect(transaction_hash).toBeTruthy();
       });